Fold Color 2.6 helpers into proto.h
One extra header for a single type code did not match how the rest of the protocol layer is laid out.
